fix(integration): fixes integration and mongo errors

+++ services/task-service/services/task_service.py
@@ -3,8 +3,9 @@ from models.task import TaskCategory, TaskItem, TaskCategoryCreate, TaskItemCrea
 from bson import ObjectId
 from datetime import datetime
 
-async def create_category(category: TaskCategoryCreate):
+async def create_category(user_id: str, category: TaskCategoryCreate):
     category_dict = {
+        "user_id": user_id,
         "name": category.name,
         "tasks": [],
         "created_at": datetime.utcnow(),
@@ -14,20 +15,20 @@ async def create_category(category: TaskCategoryCreate):
     created_category = await task_collection.find_one({"_id": result.inserted_id})
     return TaskCategory(**created_category)
 
-async def get_all_categories():
+async def get_all_categories(user_id: str):
     categories = []
-    cursor = task_collection.find()
+    cursor = task_collection.find({"user_id": user_id})
     async for document in cursor:
         categories.append(TaskCategory(**document))
     return categories
 
-async def get_category(category_id: str):
-    category = await task_collection.find_one({"_id": ObjectId(category_id)})
+async def get_category(user_id: str, category_id: str):
+    category = await task_collection.find_one({"_id": ObjectId(category_id), "user_id": user_id})
     if category:
         return TaskCategory(**category)
     return None
 
-async def add_task_to_category(category_id: str, task: TaskItemCreate):
+async def add_task_to_category(user_id: str, category_id: str, task: TaskItemCreate):
     task_dict = {
         "_id": ObjectId(),
         **task.model_dump(),
@@ -36,21 +37,22 @@ async def add_task_to_category(category_id: str, task: TaskItemCreate):
     }
     
     result = await task_collection.update_one(
-        {"_id": ObjectId(category_id)},
+        { "_id": ObjectId(category_id), "user_id": user_id },
         {
             "$push": {"tasks": task_dict},
             "$set": {"updated_at": datetime.utcnow()}
         }
     )
     if result.modified_count:
-        return await get_category(category_id)
+        return await get_category(user_id, category_id)
     return None
 
-async def update_task_status(category_id: str, task_id: str, new_status: str):
+async def update_task_status(user_id: str, category_id: str, task_id: str, new_status: str):
     result = await task_collection.update_one(
         {
             "_id": ObjectId(category_id),
-            "tasks._id": ObjectId(task_id)
+            "tasks._id": ObjectId(task_id),
+            "user_id": user_id
         },
         {
             "$set": {
@@ -61,21 +63,21 @@ async def update_task_status(category_id: str, task_id: str, new_status: str):
         }
     )
     if result.modified_count:
-        return await get_category(category_id)
+        return await get_category(user_id, category_id)
     return None
 
-async def delete_task(category_id: str, task_id: str):
+async def delete_task(user_id: str, category_id: str, task_id: str):
     result = await task_collection.update_one(
-        {"_id": ObjectId(category_id)},
+        { "_id": ObjectId(category_id), "user_id": user_id },
         {
             "$pull": {"tasks": {"_id": ObjectId(task_id)}},
             "$set": {"updated_at": datetime.utcnow()}
         }
     )
     if result.modified_count:
-        return await get_category(category_id)
+        return await get_category(user_id, category_id)
     return None
 
-async def delete_category(category_id: str):
-    result = await task_collection.delete_one({"_id": ObjectId(category_id)})
+async def delete_category(user_id: str, category_id: str):
+    result = await task_collection.delete_one({"_id": ObjectId(category_id), "user_id": user_id })
     return result.deleted_count > 0
